+++ Libraries/LibWeb/Painting/Paintable.h
@@ -92,7 +92,7 @@ public:
 
     virtual bool forms_unconnected_subtree() const { return false; }
 
-    virtual bool handle_mousewheel(Badge<EventHandler>, CSSPixelPoint, unsigned buttons, unsigned modifiers, int wheel_delta_x, int wheel_delta_y);
+    virtual bool handle_mousewheel(Badge<EventHandler>, CSSPixelPoint, unsigned buttons, unsigned modifiers, double wheel_delta_x, double wheel_delta_y);
 
     Layout::Node const& layout_node() const
     {
